Rename axios.ts to api.ts
