+++ frontend/src/api/compareAPI.ts
@@ -1,9 +1,9 @@
-import axios from 'axios';
-
 import type { SimulationOut } from '@/types/index';
 
+import api from './api';
+
 export const fetchAISimAnalysis = async (simulations: SimulationOut[]): Promise<string> => {
-  const response = await axios.post(`/analyze-simulations`, {
+  const response = await api.post(`/analyze-simulations`, {
     simulations,
   });
   return response.data.summary;
